Change HDF5 to write contig and organism table
